+++ src/bin/main.js
@@ -22,8 +22,9 @@ const runCli = async function() {
 
 const runMain = async function({ versionRange, command, args, opts }) {
   const { childProcess } = await runVersion(versionRange, command, args, {
+    ...DEFAULT_CLI_OPTS,
     ...opts,
-    ...CLI_OPTS,
+    spawn: { ...DEFAULT_CLI_OPTS.spawn, ...opts.spawn },
   })
 
   try {
@@ -34,6 +35,6 @@ const runMain = async function({ versionRange, command, args, opts }) {
   }
 }
 
-const CLI_OPTS = { spawn: { stdio: 'inherit', buffer: false } }
+const DEFAULT_CLI_OPTS = { spawn: { stdio: 'inherit', buffer: false } }
 
 runCli()